Preserve block comment formatting, add space/comment tests
We don't urgently need to trim the extra line spacing of block comment
formatting because ideally, it will already be normalized by a Markdown
renderer in the language server anyway before reaching the user's
editor.

Still, it would be nice to figure out how to do this at a later date, so
perhaps the AST can more intelligently and accurately reproduce the
original source code, including indentation and comment type. But for
the moment, this is good enough for an MVP.