Fix bug in Tokens::to_span() and map_partial_spanned()
This bug occurs when `ToSpan::to_span()` is called on an empty
zero-length `Tokens`, which produces an empty `Span` of (0, 0]. The code
in question assumes that the `Tokens` always contains at least a
`Token::Eof`, which isn't always the case.

In the case of the `map_partial_spanned()` which triggered this bug, we
now check whether the remainding tokens is empty or not before calling
`to_span()`. It is possible that `Tokens` could be redesigned to avoid
this issue entirely.