Filter unexpected tokens in lexer before checking file emptiness
Currently, the lexer only emits a "nix expressions must resolve to a
value" error if the source file is either completely empty or consists
entirely of comments.

This is fine in theory, but because the lexer currently strips invalid
tokens only _after_ the emptiness of the file is determined, files
consisting entirely of unrecognized tokens will be reduced to a
zero-length string, which doesn't trigger the nice error message and
instead causes the parser to spew quite a lot of `nom` errors.