Change nom errors to severity "bug"
This was previously the case, but was changed to severity "error"
because there was no easy way to recover and introduce custom errors in
the old parser. However, now that lexing and parsing are clearly
separated, and lexing should never fail, we are free to write our own
parser combinators which use the `nix-parser` error format natively.

As such, we change back the severity of `nom` errors to "bug" because
they are no longer meant to be seen by the end user, in favor of
clearer, more helpful, and carefully crafted error messages via `Error`.