Emit unexpected token error when alt() fails to parse
Previously, if all three `bind()` parser cases failed, `nom` would puke
out several ugly looking `ErrorKind::Alt` errors. This occurrence is
rare, since the `bind::simple()` parser is meant to be a kind of
catch-all for errors, but there exists a case where even that parser can
fail irrecoverably: failing to parse an attribute name. With this
commit, we now emit an unexpected token error in that case, just so the
diagnostics are a bit clearer and nicer looking.