Remove unused partial parser combinators
The `expect()` combinator never worked as expected (it unintentionally
screwed with `alt()` because it swallowed errors), so it should be fine
to remove.

The functionality of `skip_if_err()` is already baked into the
`many_till_partial()` and `separated_list_partial()` combinators with
the `term` argument, so it should be fine to remove as well.