Tweak atomic expr precedence to match upstream
Now that we have a separate lexer and parser, we can tweak the
precedence for atomic expressions to better match the ordering used in
the official Nix parser (see [parser.y] for details).

[parser.y]: https://github.com/NixOS/nix/blob/64ec087f582cec33733f4102ab42a3e8f291758a/src/libexpr/parser.y

This change should not impact performance, albeit with a very slight
positive effect, as shown in the benchmark results below:

```
parse_example/parse example.nix
                        time:   [659.71 us 662.17 us 665.23 us]
                        thrpt:  [735.48 KiB/s 738.87 KiB/s 741.62 KiB/s]
                 change:
                        time:   [-4.2747% -0.5544% +3.7149%] (p = 0.78 > 0.05)
                        thrpt:  [-3.5818% +0.5575% +4.4655%]
                        No change in performance detected.
Found 15 outliers among 100 measurements (15.00%)
  6 (6.00%) high mild
  9 (9.00%) high severe
```

With that said, this precedence re-ordering should provide a noticeable
performance boost in the future if we can work to reduce unnecessary
heap allocations in each parser combinator's failure path.