Change Partial::errors() to return ref to avoid clone
This method originally returned a cloned `Errors` because the `Extend`
and `FromIterator` implementations required it. However, if we access
the private field of the `Partial`, we can take ownership of it and
avoid a potentially expensive clone. There is also no need for the
`has_errors()` check anymore either.

This doesn't appear to have a significant effect on the performance of
the `example.nix` benchmark, but I imagine it could have a greater
positive impact on very large files with many binds and formal
functions which may rely heavily on `FromIterator`.

```
parse_example/parse example.nix
                        time:   [360.27 us 361.93 us 363.93 us]
                        thrpt:  [1.3129 MiB/s 1.3201 MiB/s 1.3262 MiB/s]
                 change:
                        time:   [-1.5815% +5.6174% +15.643%] (p = 0.21 > 0.05)
                        thrpt:  [-13.527% -5.3186% +1.6069%]
                        No change in performance detected.
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) high mild
  14 (14.00%) high severe
```