Avoid running partial parser twice in expect_terminated()
Currently, if the terminator parser fails, we re-trigger the partial
parser `f` again without needing to because the result is discarded by
`terminated()`. This commit changes `expect_terminated()` to cache the
result of `f` from the first run to avoid having to run it again if the
terminator is missing.