Fix line splitting for strings, add multi-line escape
The previous implementation for `split_lines_without_indentation()` only
worked correctly if the string was made up only one
`StringFragment::Literal`. The indentation level for the string was
determined by calling `input.get_column()`, which meant that every time
a new `StringFragment::Literal` was pushed into the `Token::String`, the
indentation level would be recomputed anew instead of using the
indentation of the first `StringFragment::Literal`.

This is fine for block comments and simple multi-line strings, but
strings with interpolations are rendered improperly. For example, one
would expect this:

```
''
  foo ${bar} baz
''
```

to produce the following string fragments:

```
["", "foo ", " baz", ""]
```

But instead, it produces this:

```
["", "foo ", "baz", ""]
```

Notice how `"foo "` and `" baz"` are mistakenly treated as separate
lines and the leading space for `" baz"` got trimmed, even though both
reside on the same line but different columns.

This implementation fixes the problem by adding a second parameter named
`indent` which allows the indentation level to be specified explicitly.
In the case of multi-line comments, which also happens to rely on this
function, it can simply use the previous behavior of
`split_lines_without_indent(input, input.get_column())`. But multi-line
strings can use the second argument to specify the correct indentation
level when lexing multiple `StringFragment::Literal`s.

Another addition in this commit was the ability to escape the `$`
character in multi-line strings with a preceding `''`. This was tricky
to implement, but appears to work as expected. The line splitting bug
described above was discovered while attempting to implement this
feature.