Use smallvec to reduce heap allocations in parser
There is a common case in `nix-parser` where only one error is pushed
onto the `Errors` stack, usually an `Error::Nom` while parsing tokens
which signify the beginning of a new language construct, e.g. `if`
marking the beginning of an `if` statement, or `{` marking the beginning
of a set or formal function declaration.

There are many areas in the parser which use `nom::combinator::alt` to
test a variety of different branches in quick succession and choose the
first one that advances past the first token. Unfortunately, since the
`Errors` stack is backed by a `Vec`, each case being attempted by `alt`
will allocate the error on the heap and destroy the allocation almost
immediately after.

To fix this wastefulness, we replace the `Vec` inside the `Errors`
stack with a [`SmallVec`] and allow the first `push()` call to allocate
"for free" on the stack instead of the heap. Any additional pushes to
`Errors` will subsequently be placed on the heap like before. This is a
reasonable tradeoff because it is very common for the first `push()` to
be an `Error::Nom` which is lightweight and doesn't require much space.

[`SmallVec`]: https://docs.rs/smallvec/0.6.12/smallvec/struct.SmallVec.html

Switching the backing store of `Errors` to a `SmallVec<[Error; 1]>`
provides a noticeable boost to performance, finally breaking through the
2 MiB/s throughput threshold in the benchmark below:

```
parse_example/parse example.nix
                        time:   [212.84 us 214.75 us 217.07 us]
                        thrpt:  [2.2011 MiB/s 2.2249 MiB/s 2.2448 MiB/s]
                 change:
                        time:   [-19.280% -14.825% -9.8772%] (p = 0.00 < 0.05)
                        thrpt:  [+10.960% +17.405% +23.885%]
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe
```