Store custom error messages as copy-on-write strings
This should reduce some heap allocations when processing code with
a large number of mistakes. This commit ensures that statically embedded
error messages in `Error::Message` aren't allocated on the heap. The
`example.rs` Criterion benchmark did not detect any change in
performance because `Error::Message` is never used in the hot path.