Use copy-on-write strings for all error types
Surprisingly, this simple change provides a quite large performance
improvement for the hot path, as shown by the impressive benchmark
results below:

```
parse_example/parse example.nix
                        time:   [197.90 us 198.83 us 199.96 us]
                        thrpt:  [2.3894 MiB/s 2.4030 MiB/s 2.4143 MiB/s]
                 change:
                        time:   [-17.367% -11.301% -5.4713%] (p = 0.00 < 0.05)
                        thrpt:  [+5.7880% +12.740% +21.017%]
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) high mild
  10 (10.00%) high severe
```

Even though these errors are not displayed to the user, they appear to
have a significant impact on performance. It is suspected to be a side
effect of the ambiguity between sets and formal functions (see issue #6
and PR #8), where the Nix language grammar requires significant
backtracking when parsing any construct beginning with `{`.

Perhaps there exist some additional optimizations that could eliminate
the construction of these errors altogether, but for now, this small
change suffices as an easy win.