Retain parse error history only in debug builds
This commit disables the retention of `nom` parse errors in the `Errors`
struct in release builds, opting to only keep the last error to occur,
if any. This is similar in behavior to the `(I, ErrorKind)` error type
used by default in simple `nom` parsers.

Disabling this history retention allows us to keep the `SmallVec`-backed
error stack small and avoid spilling over onto the heap when we don't
need to, and it results in a tidy performance improvement in the
benchmark below:

```
parse_example/parse example.nix
                        time:   [188.32 us 189.92 us 191.64 us]
                        thrpt:  [2.4932 MiB/s 2.5158 MiB/s 2.5371 MiB/s]
                 change:
                        time:   [-13.425% -7.0828% -1.3051%] (p = 0.02 < 0.05)
                        thrpt:  [+1.3224% +7.6227% +15.507%]
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) high mild
  7 (7.00%) high severe
```

Since `Error::Nom` is never supposed to be displayed to the user and is
only displayed in case of an unrecoverable, catastrophic bug in the
parser, we don't need to carry the entire error trace around. Still, not
having the complete trace makes debugging significantly harder, so this
change is limited to release builds only.