Add nix-parser2 crate with simplified lexer
This commit lays the groundwork for a forthcoming redesign of the Nix
parser. This new parser will produce a lossless CST (concrete syntax
tree) as an intermediate step towards producing an AST. This CST should
be better suited for uses beyond expression parsing and evaluation, such
as formatting, completion, and symbol lookup because it retains all
whitespace, comments, and unknown tokens in the resulting token stream.

The new `nix-parser2` crate introduces a new lexer with the following
key changes from the original `nix-parser` crate:

* Lazy: no nested collections of interpolations and string fragments.
* Lossless: all spans are preserved, no more strings in certain tokens.
* Snapshot tests: uses `serde_json` to de/serialize token streams.
* Basic validation: detects unclosed block comments or trailing slashes
  in path literals without offloading extra work on the parser.

Note that this lexer doesn't yet support string interplations nor
delimiter balancing, so there is still work to be done.

This lexer continues to rely on `nom` and `nom_locate` for two reasons:

1. More declarative-looking and easier to read than a hand-rolled lexer.
2. Lexer does not need to be incremental, nor does it require a complex
   `Partial` abstraction like the current parser does.

However, the upcoming parser will not use `nom` and will be hand-rolled
instead. Writing fault-tolerant parsers that emit good diagnostics with
`nom` is very difficult and results in very messy and tangled code (see
`partial.rs`), and string interning is more difficult to pull off due to
each parser being a pure function.