Fix lexer mode pushing/popping logic in new lexer
This commit changes the lexer mode pushing/popping and brace counting
logic to better match what the upstream [Nix lexer] does. We also
increase the number of inline `SmallVec` elements to 8, since we are now
using it for counting braces and we would like to avoid spilling over
onto the heap for deeply nested Nix expressions.

[Nix lexer]: https://github.com/NixOS/nix/blob/master/src/libexpr/lexer.l