Define separate ToSpan trait in nix-parser2
This is because we wish to implement `ToSpan` on several external types,
e.g. `rowan::TextRange` and `rowan::SyntaxNode`, and we cannot do this
in the `nix-parser2` crate if the trait is defined in the `nix-lexer`
crate. As such, we make `nix_lexer::ToSpan` private and redefine it in
`nix-parser2` instead.